refactor : 당첨 번호 일치 계산 테스트 클래스 이동
