+++ src/test/java/lotto/domain/LottoTest.java
@@ -3,6 +3,9 @@ package lotto.domain;
 import lotto.exception.ErrorMessage;
 import org.junit.jupiter.api.DisplayName;
 import org.junit.jupiter.api.Test;
+import org.junit.jupiter.params.ParameterizedTest;
+import org.junit.jupiter.params.provider.Arguments;
+import org.junit.jupiter.params.provider.MethodSource;
 
 import java.util.Arrays;
 import java.util.List;
@@ -10,9 +13,24 @@ import java.util.stream.Collectors;
 import java.util.stream.IntStream;
 import java.util.stream.Stream;
 
+import static org.assertj.core.api.Assertions.assertThat;
 import static org.assertj.core.api.Assertions.assertThatIllegalArgumentException;
+import static org.junit.jupiter.params.provider.Arguments.arguments;
 
 class LottoTest {
+
+    static Stream<Arguments> matchCalculationParametersProvider() {
+        List<LottoNumber> lottoNumbers = Stream.of(1, 2, 3, 4, 5, 6)
+                .map(LottoNumber::new)
+                .collect(Collectors.toList());
+
+        return Stream.of(
+                arguments(lottoNumbers, Arrays.asList(1, 2, 3, 11, 22, 33), 3),
+                arguments(lottoNumbers, Arrays.asList(1, 2, 3, 4, 11, 22), 4),
+                arguments(lottoNumbers, Arrays.asList(1, 2, 3, 4, 5, 11), 5),
+                arguments(lottoNumbers, Arrays.asList(1, 2, 3, 4, 5, 6), 6)
+        );
+    }
 
     @DisplayName("로또 번호가 6개가 아닐 시 예외")
     @Test
@@ -45,5 +63,19 @@ class LottoTest {
         assertThatIllegalArgumentException().isThrownBy(() -> {
             new LottoNumber(46);
         }).withMessage(ErrorMessage.LOTTO_NUMBER_RANGE.getMessage());
+    }
+
+    @DisplayName("당첨 번호 일치 계산")
+    @ParameterizedTest
+    @MethodSource("matchCalculationParametersProvider")
+    void matchCalculation(List<LottoNumber> lottoNumber, List<Integer> winningNumber, int winningNumberMatchCount) {
+        //given
+        Lotto lotto = new Lotto(lottoNumber);
+
+        //when
+        int winningNumberMatchResult = lotto.winningNumberMatchCount(winningNumber);
+
+        //then
+        assertThat(winningNumberMatchResult).isEqualTo(winningNumberMatchCount);
     }
 }
