+++ src/test/java/lotto/domain/WinningTest.java
@@ -16,33 +16,6 @@ import static org.junit.jupiter.params.provider.Arguments.arguments;
 
 class WinningTest {
 
-    static Stream<Arguments> matchCalculationParametersProvider() {
-        List<LottoNumber> lottoNumbers = Stream.of(1, 2, 3, 4, 5, 6)
-                .map(LottoNumber::new)
-                .collect(Collectors.toList());
-
-        return Stream.of(
-                arguments(lottoNumbers, Arrays.asList(1, 2, 3, 11, 22, 33), 3),
-                arguments(lottoNumbers, Arrays.asList(1, 2, 3, 4, 11, 22), 4),
-                arguments(lottoNumbers, Arrays.asList(1, 2, 3, 4, 5, 11), 5),
-                arguments(lottoNumbers, Arrays.asList(1, 2, 3, 4, 5, 6), 6)
-        );
-    }
-
-    @DisplayName("당첨 번호 일치 계산")
-    @ParameterizedTest
-    @MethodSource("matchCalculationParametersProvider")
-    void matchCalculation(List<LottoNumber> lottoNumber, List<Integer> winningNumber, int winningNumberMatchCount) {
-        //given
-        Lotto lotto = new Lotto(lottoNumber);
-
-        //when
-        int winningNumberMatchResult = lotto.winningNumberMatchCount(winningNumber);
-
-        //then
-        assertThat(winningNumberMatchResult).isEqualTo(winningNumberMatchCount);
-    }
-
     @DisplayName("당첨 결과 통계")
     @Test
     void winningResultTest() {
